dod.ctd.bbmp() drops the 'read' parameter
I'm in the middle of other changes and am just pushing these to have a
clean commit history.  I'm not even sure the package will build/test
with this exact commit.  Later today, I'll alter ctd.bats.R and ensure
that any tests are also updated and then I'll push the final (then
working) version.

All of this is pursuant to a Z with JH earlier today.