alterations to 'check' works...
Check was failing because of local files that had been downloaded.  I
made the dod.met.sounding() example use a tempdir, and I think we ought
to do that for every test, every doc page, and the vignettes.  But
sometimes it's OK.  I don't know whether the problem is actually in
saving files or in my confusion as to how to name such files in
.Rbuildignore, but I am really quite sure that CRAN will reject a
package that downloads to the user's directory so we ought to convert
any examples/test/vignettes that download data so that destdir is used,
and cleaned up afterwards.  I just don't feel like doing that at the
moment, and it builds+tests+checks OK at this moment so I'm leaving the
issue.