I think this will fix the webpage because...
'docs' was in .gitignore.  Perhaps that is because we used to do remote
building of the docs.  I dropped that method because there were problems
and I just didn't see the point in debugging yaga [yet another github
action] that might be caused by remote server problems, by changes in
requisite actions (some actions seem to get a new version every year or
two).